Split Interface into Ethernet and IP parts.
- State, Config and Processor all get split in Ethernet and IP counterparts.
- Functions that only deal with IP packets are moved to the IP module, mostly
  as-is. No changes to the logic have been done, only changes to get the code to build.
- 2 new traits are introduced:
    - LowerDispatcher (for outgoing packets)
    - LowerProcessor (for incoming packets).
  The IP module uses them to call back into the Ethernet module, so it can handle
  Ethernet-specific concerns (arp, ndisc) and send packets to the Device.
- Ethernet's process_ipv{4,6} deal with the Ethernet-specific concerns and then call their IP counterparts.
